visualisations as seperate scripts (KISS)
