+++ experiments/depth_first/depth_first_experiment.json
@@ -1,6 +1,5 @@
 {
     "data_folder": "nl",
-    "scheme": 1,
     "runs": [
         {
             "function": "depth_first",
