+++ pyproject.toml
@@ -6,8 +6,9 @@ readme = "README.md"
 requires-python = ">=3.13"
 dependencies = [
     "colour==0.1.5",
-    "bokeh==3.0.3",
-    "matplotlib==3.6.3",
+    "bokeh==3.9.1",
+    "matplotlib==3.11.0",
+    "numpy==2.5.0",
 ]
 
 [build-system]
