+++ visualisation/greedy/base_vs_random_graph.py
@@ -1,24 +1,6 @@
 import matplotlib.pyplot as plt
 import csv
 
-__all__ = [
-    "random_greedy_graph",
-    "base_vs_random_graph",
-]
-
-
-def random_greedy_graph() -> None:
-    fig, ax = plt.subplots()
-    with open("results/greedy/random_greedy.csv", 'r') as input_file:
-        result_reader = csv.reader(input_file, delimiter=',')
-        results = [int(value) for value, _ in result_reader]
-
-    ax.hist(results)
-    ax.set_title('Random Greedy Solutions')
-    ax.set_xlabel('Total Costs')
-    ax.set_ylabel('Number of Solutions')
-    fig.savefig("results/greedy/random_greedy.png")
-
 
 def base_vs_random_graph() -> None:
     fig, ax = plt.subplots()
@@ -38,3 +20,7 @@ def base_vs_random_graph() -> None:
     ax.set_xlabel('Total Costs')
     ax.set_ylabel('Number of Solutions')
     fig.savefig("results/greedy/base_vs_random.png")
+
+
+if __name__ == "__main__":
+    base_vs_random_graph()
